Changed script name from formatname.py to addnewsequences.py. That module also adds the matching sequences of another file to the EcOH database.
